requirements.txt
flask
selenium
requests
python-dotenv
